tools/requirements.txt
# ESP32控制板CAN工具依赖包
# 
# 安装方法:
#   pip install -r requirements.txt
#
# 或者使用国内镜像源:
#   pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple/

# 核心CAN通信库
python-can>=4.0.0

# 串口通信支持（用于串口CAN设备）
pyserial>=3.5

# PEAK CAN支持（可选，如果使用PEAK CAN设备）
# python-can[pcan]>=4.0.0

# Vector CAN支持（可选，如果使用Vector CAN设备）
# python-can[vector]>=4.0.0

# Kvaser CAN支持（可选，如果使用Kvaser CAN设备）
# python-can[kvaser]>=4.0.0

# USB2CAN支持（可选）
# python-can[usb2can]>=4.0.0

# 其他实用工具
colorama>=0.4.4  # 彩色终端输出
